extract mettagrid.map.utils.storage code; handle file:// uris
